requirements.txt
# ===========================================================================
# 3-Way PO Reconciliation Platform -- Production Dependencies
# Generated: 2026-03-31  (pinned to currently installed versions)
# ===========================================================================

# ---------------------------------------------------------------------------
# Core Django stack
# ---------------------------------------------------------------------------
Django==5.0.14
djangorestframework==3.16.1
django-filter==25.1
django_celery_results==2.6.0
django-import-export==4.3.12

# ---------------------------------------------------------------------------
# Database
# ---------------------------------------------------------------------------
mysqlclient==2.2.8

# ---------------------------------------------------------------------------
# Task queue
# ---------------------------------------------------------------------------
celery==5.6.2
redis==6.4.0

# ---------------------------------------------------------------------------
# Fuzzy / string matching
# ---------------------------------------------------------------------------
thefuzz==0.22.1
python-Levenshtein==0.27.3
RapidFuzz==3.14.3

# ---------------------------------------------------------------------------
# Utilities
# ---------------------------------------------------------------------------
python-dotenv==1.2.2
dateparser==1.3.0
requests==2.31.0
beautifulsoup4==4.14.3
playwright==1.58.0
pydantic==2.12.5
openpyxl==3.1.2
reportlab>=4.0
Pillow==12.1.1

# ---------------------------------------------------------------------------
# AI / LLM
# ---------------------------------------------------------------------------
openai==2.30.0
langchain-openai==1.1.12
langfuse==4.0.1

# ---------------------------------------------------------------------------
# Azure services
# ---------------------------------------------------------------------------
azure-ai-formrecognizer==3.3.2
azure-core==1.34.0
azure-storage-blob==12.20.0

# ---------------------------------------------------------------------------
# Observability
# ---------------------------------------------------------------------------
opentelemetry-api==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-exporter-otlp-proto-http==1.40.0
openinference-instrumentation==0.1.46
openinference-instrumentation-openai==0.1.43
openinference-semantic-conventions==0.1.28
python-json-logger==2.0.7
python-logging-loki==0.3.1
